fixed type warnings

+++ hw_2/Task4/main.c
@@ -6,23 +6,23 @@ int main(int argc, char* argv[])
     puts("A -> B = !A || B");
     puts("A\tB\t!A\t!A || B");
 
-    for (size_t i = 0; i <= 1; i++)
+    for (unsigned int i = 0; i <= 1; i++)
     {
-        for (size_t j = 0; j <= 1; j++)
+        for (unsigned int j = 0; j <= 1; j++)
         {
-            printf("%d\t%d\t%d\t%d\n", i, j, !i, !i || j);
+            printf("%u\t%u\t%u\t%u\n", i, j, !i, !i || j);
         }
-    }
+    }   
 
     puts("\nA <-> B = (A && B) || (!A && !B)");
     puts("A\tB\tA && B\t!A\t!B\t(!A && !B)\t(A && B) || (!A && !B)");
 
-    for (size_t i = 0; i <= 1; i++)
+    for (unsigned int i = 0; i <= 1; i++)
     {
-        for (size_t j = 0; j <= 1; j++)
+        for (unsigned int j = 0; j <= 1; j++)
         {
-            printf("%d\t%d\t%d\t%d\t%d\t%d\t\t%d\n",
-                i, j, i && j, !i, !j, ! i&& !j, i && j || !i && !j
+            printf("%u\t%u\t%u\t%u\t%u\t%u\t\t%u\n",
+                i, j, i && j, !i, !j, ! i&& !j, (i && j) || (!i && !j)
             );
         }
     }
